+++ packages/analytics/src/AnalyticsContext.ts
@@ -11,11 +11,12 @@ export interface AnalyticsEvents {
     customerSuggestionExecute(): void;
     customerPaymentMethodExecuted(payload?: CheckoutPaymentMethodExecutedOptions): void;
     showShippingMethods(): void;
-    selectedPaymentMethod(methodName?: string): void;
+    selectedPaymentMethod(methodName: string, methodId: string): void;
     clickPayButton(payload?: { [key: string]: unknown }): void;
     paymentRejected(): void;
     paymentComplete(): void;
     exitCheckout(): void;
+    walletButtonClick(methodId: string): void;
 }
 
 export interface AnalyticsContextProps {