+++ packages/analytics/src/AnalyticsProvider.mock.tsx
@@ -23,6 +23,7 @@ const AnalyticsProviderMock = ({ children, analyticsTracker = {} }: AnalyticsPro
         paymentRejected: jest.fn(),
         paymentComplete: jest.fn(),
         exitCheckout: jest.fn(),
+        walletButtonClick: jest.fn(),
     };
 
     return (