+++ packages/checkout-button-integration/src/CheckoutButton.tsx
@@ -11,6 +11,7 @@ const CheckoutButton: FunctionComponent<CheckoutButtonProps> = ({
     containerId,
     methodId,
     onUnhandledError,
+    onWalletButtonClick,
 }) => {
     useEffect(() => {
         initializeCustomer({
@@ -18,13 +19,21 @@ const CheckoutButton: FunctionComponent<CheckoutButtonProps> = ({
             [methodId]: {
                 container: containerId,
                 onUnhandledError,
+                onClick: () => onWalletButtonClick(methodId),
             },
         }).catch(onUnhandledError);
 
         return () => {
             deinitializeCustomer({ methodId }).catch(onUnhandledError);
         };
-    }, [containerId, deinitializeCustomer, initializeCustomer, methodId, onUnhandledError]);
+    }, [
+        containerId,
+        deinitializeCustomer,
+        initializeCustomer,
+        methodId,
+        onUnhandledError,
+        onWalletButtonClick,
+    ]);
 
     return <div id={containerId} />;
 };