+++ packages/core/src/app/customer/CheckoutButtonContainer.tsx
@@ -1,6 +1,6 @@
 import { CheckoutSelectors, CheckoutService } from '@bigcommerce/checkout-sdk';
 import classNames from 'classnames';
-import React, { FunctionComponent } from 'react';
+import React, { FunctionComponent, memo } from 'react';
 
 import { TranslatedString, useLocale } from '@bigcommerce/checkout/locale';
 import { CheckoutContextProps } from '@bigcommerce/checkout/payment-integration-api';
@@ -17,6 +17,7 @@ interface CheckoutButtonContainerProps {
     isPaymentStepActive: boolean;
     checkEmbeddedSupport(methodIds: string[]): void;
     onUnhandledError(error: Error): void;
+    onWalletButtonClick(methodId: string): void;
 }
 
 interface WithCheckoutCheckoutButtonContainerProps {
@@ -33,19 +34,6 @@ const paypalCommerceIds = [
     'paypalcommercevenmo',
 ];
 
-const sortMethodIds = (methodIds:string[]): string[] => {
-    const order = [
-        'applepay',
-        'braintreepaypalcredit',
-        'braintreepaypal',
-        'paypalcommercevenmo',
-        'paypalcommercecredit',
-        'paypalcommerce',
-    ];
-
-    return methodIds.sort((a, b) => order.indexOf(b) - order.indexOf(a));
-}
-
 const isPayPalCommerce = (methodId: string): boolean => paypalCommerceIds.includes(methodId);
 
 const CheckoutButtonContainer: FunctionComponent<CheckoutButtonContainerProps & WithCheckoutCheckoutButtonContainerProps> = (
@@ -58,6 +46,7 @@ const CheckoutButtonContainer: FunctionComponent<CheckoutButtonContainerProps & 
         isPaymentStepActive,
         initializedMethodIds,
         onUnhandledError,
+        onWalletButtonClick,
     }) => {
     const { language } = useLocale();
 
@@ -84,6 +73,7 @@ const CheckoutButtonContainer: FunctionComponent<CheckoutButtonContainerProps & 
                 key={methodId}
                 methodId={methodId}
                 onError={onUnhandledError}
+                onClick={onWalletButtonClick}
             />
         }
 
@@ -95,6 +85,7 @@ const CheckoutButtonContainer: FunctionComponent<CheckoutButtonContainerProps & 
                     language={language}
                     methodId={methodId}
                     onUnhandledError={onUnhandledError}
+                    onWalletButtonClick={onWalletButtonClick}
                 />;
     });
 
@@ -161,10 +152,10 @@ function mapToCheckoutButtonContainerProps({
     return {
         checkoutService,
         checkoutState,
-        availableMethodIds: sortMethodIds(availableMethodIds),
+        availableMethodIds,
         initializedMethodIds,
         isLoading,
     }
 }
 
-export default withCheckout(mapToCheckoutButtonContainerProps)(CheckoutButtonContainer);
+export default memo(withCheckout(mapToCheckoutButtonContainerProps)(CheckoutButtonContainer));