+++ packages/core/src/app/customer/CheckoutButtonList.tsx
@@ -45,6 +45,7 @@ export interface CheckoutButtonListProps {
     deinitialize(options: CustomerRequestOptions): void;
     initialize(options: CustomerInitializeOptions): void;
     onError?(error: Error): void;
+    onClick?(methodId: string): void;
 }
 
 export const filterUnsupportedMethodIds = (methodIds:string[]): string[] => {