+++ packages/test-framework/src/fixture/pagePreset/index.ts
@@ -4,4 +4,5 @@
 export { CheckoutPagePreset } from './CheckoutPagePreset';
 export { CustomerStepAsGuest as CustomerStepPreset } from './CustomerStepAsGuest';
 export { PaymentStepAsGuest as PaymentStepAsGuestPreset } from './PaymentStepAsGuest';
+export { PaymentStepAsGuestEU as PaymentStepAsGuestEUPreset } from './PaymentStepAsGuestEU';
 export { UseAUD as UseAUDPreset } from './UseAUD';